# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nabiha Financial Intelligence Platform API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type Currency = (typeof Currency)[keyof typeof Currency];

export const Currency = {
  JOD: "JOD",
  SAR: "SAR",
  AED: "AED",
  EGP: "EGP",
  KWD: "KWD",
  BHD: "BHD",
  QAR: "QAR",
  OMR: "OMR",
  IQD: "IQD",
  LBP: "LBP",
  MAD: "MAD",
  TND: "TND",
  DZD: "DZD",
  LYD: "LYD",
  YER: "YER",
  SDG: "SDG",
  SYP: "SYP",
  USD: "USD",
} as const;

export type Priority = (typeof Priority)[keyof typeof Priority];

export const Priority = {
  essential: "essential",
  important: "important",
  luxury: "luxury",
} as const;

export interface UserProfile {
  userId: string;
  monthlySalary: number;
  currency: Currency;
  onboardingComplete: boolean;
  displayName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateProfileBody {
  monthlySalary?: number;
  currency?: Currency;
  onboardingComplete?: boolean;
  displayName?: string | null;
}

export interface Expense {
  id: number;
  title: string;
  amount: number;
  priority: Priority;
  categoryId?: number | null;
  categoryName?: string | null;
  notes?: string | null;
  date: string;
  createdAt: string;
}

export interface CreateExpenseBody {
  title: string;
  amount: number;
  priority: Priority;
  categoryId?: number | null;
  notes?: string | null;
  date: string;
}

export interface UpdateExpenseBody {
  title?: string;
  amount?: number;
  priority?: Priority;
  categoryId?: number | null;
  notes?: string | null;
  date?: string;
}

export interface Category {
  id: number;
  name: string;
  icon?: string | null;
  color?: string | null;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  icon?: string | null;
  color?: string | null;
}

export interface Commitment {
  id: number;
  title: string;
  amount: number;
  dueDay: number;
  isPaid: boolean;
  notes?: string | null;
  createdAt: string;
}

export interface CreateCommitmentBody {
  title: string;
  amount: number;
  dueDay: number;
  notes?: string | null;
}

export interface UpdateCommitmentBody {
  title?: string;
  amount?: number;
  dueDay?: number;
  isPaid?: boolean;
  notes?: string | null;
}

export type SubscriptionFrequency =
  (typeof SubscriptionFrequency)[keyof typeof SubscriptionFrequency];

export const SubscriptionFrequency = {
  monthly: "monthly",
  yearly: "yearly",
  weekly: "weekly",
} as const;

export type SubscriptionStatus =
  (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus];

export const SubscriptionStatus = {
  active: "active",
  inactive: "inactive",
  upcoming: "upcoming",
} as const;

export type SubscriptionCategory =
  (typeof SubscriptionCategory)[keyof typeof SubscriptionCategory];

export const SubscriptionCategory = {
  streaming: "streaming",
  music: "music",
  productivity: "productivity",
  fitness: "fitness",
  other: "other",
} as const;

export interface Subscription {
  id: number;
  name: string;
  amount: number;
  frequency: SubscriptionFrequency;
  category: SubscriptionCategory;
  nextRenewalDate: string;
  status: SubscriptionStatus;
  brandColor?: string | null;
  brandIcon?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateSubscriptionBody {
  name: string;
  amount: number;
  frequency: SubscriptionFrequency;
  category: SubscriptionCategory;
  nextRenewalDate: string;
  status?: SubscriptionStatus;
  brandColor?: string | null;
  brandIcon?: string | null;
  notes?: string | null;
}

export interface UpdateSubscriptionBody {
  name?: string;
  amount?: number;
  frequency?: SubscriptionFrequency;
  category?: SubscriptionCategory;
  nextRenewalDate?: string;
  status?: SubscriptionStatus;
  brandColor?: string | null;
  brandIcon?: string | null;
  notes?: string | null;
}

export type EventType = (typeof EventType)[keyof typeof EventType];

export const EventType = {
  salary: "salary",
  bill: "bill",
  subscription: "subscription",
  goal: "goal",
  reminder: "reminder",
  expense: "expense",
} as const;

export interface Event {
  id: number;
  title: string;
  amount?: number | null;
  date: string;
  type: EventType;
  notes?: string | null;
  createdAt: string;
}

export interface CreateEventBody {
  title: string;
  amount?: number | null;
  date: string;
  type: EventType;
  notes?: string | null;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  info: "info",
  warning: "warning",
  success: "success",
  danger: "danger",
} as const;

export interface Notification {
  id: number;
  title: string;
  message: string;
  type: NotificationType;
  isRead: boolean;
  link?: string | null;
  createdAt: string;
}

export interface SimulatorBody {
  loanAmount: number;
  /** Percentage e.g. 5.5 */
  annualInterestRate: number;
  termMonths: number;
  /** Optional override of saved profile salary */
  monthlySalary?: number | null;
}

export type SimulatorResultAffordability =
  (typeof SimulatorResultAffordability)[keyof typeof SimulatorResultAffordability];

export const SimulatorResultAffordability = {
  safe: "safe",
  caution: "caution",
  risky: "risky",
} as const;

export type SimulatorResultProjectedBalanceItem = {
  month: number;
  remainingBalance: number;
};

export interface SimulatorResult {
  monthlyPayment: number;
  totalInterest: number;
  totalRepayment: number;
  /** Percentage of monthly salary consumed by this payment */
  debtToIncomeRatio: number;
  commitmentsTotal: number;
  /** Existing commitments + this payment vs salary, percentage */
  currentDebtToIncomeRatio: number;
  affordability: SimulatorResultAffordability;
  projectedBalance: SimulatorResultProjectedBalanceItem[];
}

export interface DashboardSummary {
  totalThisMonth: number;
  totalLastMonth: number;
  essentialTotal: number;
  importantTotal: number;
  luxuryTotal: number;
  commitmentsTotal: number;
  unpaidCommitmentsCount: number;
  subscriptionsTotal: number;
  activeSubscriptionsCount: number;
  expensesCount: number;
  topCategory?: string | null;
  monthlySalary: number;
  currency: Currency;
  remainingBalance: number;
}

export interface PriorityBreakdown {
  priority: Priority;
  total: number;
  count: number;
  percentage: number;
}

export interface CategoryBreakdown {
  categoryId?: number | null;
  categoryName: string;
  total: number;
  count: number;
  percentage: number;
}

export interface MonthlyTrend {
  /** YYYY-MM format */
  month: string;
  total: number;
  essential: number;
  important: number;
  luxury: number;
}

export type SubscriptionsBreakdownByCategoryItem = {
  category: SubscriptionCategory;
  total: number;
  count: number;
};

export interface SubscriptionsBreakdown {
  monthlyTotal: number;
  yearlyTotal: number;
  byCategory: SubscriptionsBreakdownByCategoryItem[];
}

export type ListExpensesParams = {
  categoryId?: number;
  priority?: Priority;
  /**
   * YYYY-MM format
   */
  month?: string;
};

export type ListEventsParams = {
  year?: number;
  /**
   * YYYY-MM format
   */
  month?: string;
};

export type GetDashboardSummaryParams = {
  /**
   * YYYY-MM format
   */
  month?: string;
};

export type GetPriorityBreakdownParams = {
  month?: string;
};

export type GetCategoryBreakdownParams = {
  month?: string;
};