# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Nabiha Financial Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List subscriptions
      responses:
        "200":
          description: Subscriptions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Created subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    put:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionBody"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Delete a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /simulator/borrowing-capacity:
    post:
      operationId: simulateBorrowingCapacity
      tags: [simulator]
      summary: Calculate borrowing capacity given desired loan params
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulatorBody"
      responses:
        "200":
          description: Simulator result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatorResult"
  /summary/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [summary]
      summary: Get dashboard summary with totals and breakdowns
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM format"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /summary/priority-breakdown:
    get:
      operationId: getPriorityBreakdown
      tags: [summary]
      summary: Get spending breakdown by priority level
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Priority breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriorityBreakdown"
  /summary/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [summary]
      summary: Get spending breakdown by category
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
...
components:
  schemas:
...
    Currency:
      type: string
      enum: [JOD, SAR, AED, EGP, KWD, BHD, QAR, OMR, IQD, LBP, MAD, TND, DZD, LYD, YER, SDG, SYP, USD]
    Priority:
      type: string
      enum: [essential, important, luxury]
...
    SubscriptionFrequency:
      type: string
      enum: [monthly, yearly, weekly]
    SubscriptionStatus:
      type: string
      enum: [active, inactive, upcoming]
    SubscriptionCategory:
      type: string
      enum: [streaming, music, productivity, fitness, other]
    Subscription:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        amount:
          type: number
        frequency:
          $ref: "#/components/schemas/SubscriptionFrequency"
        category:
          $ref: "#/components/schemas/SubscriptionCategory"
        nextRenewalDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/SubscriptionStatus"
        brandColor:
          type: string
          nullable: true
        brandIcon:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - amount
        - frequency
        - category
        - nextRenewalDate
        - status
        - createdAt
    CreateSubscriptionBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        frequency:
          $ref: "#/components/schemas/SubscriptionFrequency"
        category:
          $ref: "#/components/schemas/SubscriptionCategory"
        nextRenewalDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/SubscriptionStatus"
        brandColor:
          type: string
          nullable: true
        brandIcon:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
        - amount
        - frequency
        - category
        - nextRenewalDate
    UpdateSubscriptionBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        frequency:
          $ref: "#/components/schemas/SubscriptionFrequency"
        category:
          $ref: "#/components/schemas/SubscriptionCategory"
        nextRenewalDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/SubscriptionStatus"
        brandColor:
          type: string
          nullable: true
        brandIcon:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    SimulatorBody:
      type: object
      properties:
        loanAmount:
          type: number
        annualInterestRate:
          type: number
          description: Percentage e.g. 5.5
        termMonths:
          type: integer
        monthlySalary:
          type: number
          nullable: true
          description: Optional override of saved profile salary
      required:
        - loanAmount
        - annualInterestRate
        - termMonths
    SimulatorResult:
      type: object
      properties:
        monthlyPayment:
          type: number
        totalInterest:
          type: number
        totalRepayment:
          type: number
        debtToIncomeRatio:
          type: number
          description: Percentage of monthly salary consumed by this payment
        commitmentsTotal:
          type: number
        currentDebtToIncomeRatio:
          type: number
          description: Existing commitments + this payment vs salary, percentage
        affordability:
          type: string
          enum: [safe, caution, risky]
        projectedBalance:
          type: array
          items:
            type: object
            properties:
              month:
                type: integer
              remainingBalance:
                type: number
            required:
              - month
              - remainingBalance
      required:
        - monthlyPayment
        - totalInterest
        - totalRepayment
        - debtToIncomeRatio
        - commitmentsTotal
        - currentDebtToIncomeRatio
        - affordability
        - projectedBalance
    DashboardSummary:
      type: object
      properties:
        totalThisMonth:
          type: number
        totalLastMonth:
          type: number
        essentialTotal:
          type: number
        importantTotal:
          type: number
        luxuryTotal:
          type: number
        commitmentsTotal:
          type: number
        unpaidCommitmentsCount:
          type: integer
        subscriptionsTotal:
          type: number
        activeSubscriptionsCount:
          type: integer
        expensesCount:
          type: integer
        topCategory:
          type: string
          nullable: true
        monthlySalary:
          type: number
        currency:
          $ref: "#/components/schemas/Currency"
        remainingBalance:
          type: number
      required:
        - totalThisMonth
        - totalLastMonth
        - essentialTotal
        - importantTotal
        - luxuryTotal
        - commitmentsTotal
        - unpaidCommitmentsCount
        - subscriptionsTotal
        - activeSubscriptionsCount
        - expensesCount
        - monthlySalary
        - currency
        - remainingBalance
    PriorityBreakdown:
      type: object
      properties:
        priority:
          $ref: "#/components/schemas/Priority"
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
      required:
        - priority
        - total
        - count
        - percentage
    CategoryBreakdown:
      type: object
      properties:
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
      required:
        - categoryName
        - total
        - count
        - percentage
...